requirements.txt
Flask==2.0.2
jsonschema==4.0.1
numpy==1.21.2
Pillow==8.3.2
pyTelegramBotAPI==3.7.5
requests==2.26.0
thefuzz